+++ packages/plugin-react/package.json
@@ -69,6 +69,6 @@
     "@types/react-dom": "^17.0.0",
     "@types/react-router-dom": "^5.1.3",
     "@types/webpack": "^4.41.10",
-    "ssr-types": "^5.5.57"
+    "ssr-types-react": "^5.5.57"
   }
 }
