+++ packages/plugin-react/src/entry/create-context.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'react'
-import { IContext } from 'ssr-types'
+import { IContext } from 'ssr-types-react'
 
 // 服务端需要为每一个请求创建全新的 contetx 方式请求污染
 const serverContext = (context?: any) => {
