+++ packages/types-react/CHANGELOG.md
@@ -1,0 +1,128 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [5.5.57](https://github.com/ykfe/ssr/compare/v5.5.56...v5.5.57) (2021-06-23)
+
+
+### Bug Fixes
+
+* restore @types/react in ssr-types temporary ([b54cb5c](https://github.com/ykfe/ssr/commit/b54cb5ce004de3673c4d244e8a9a922505537621))
+
+
+
+
+
+## [5.5.55](https://github.com/ykfe/ssr/compare/v5.5.54...v5.5.55) (2021-06-22)
+
+
+### Bug Fixes
+
+* remove @types/react in ssr-types ([de24c71](https://github.com/ykfe/ssr/commit/de24c71be48728465c30efde89bd5f68a76fe08b))
+
+
+
+
+
+## [5.5.51](https://github.com/ykfe/ssr/compare/v5.5.50...v5.5.51) (2021-06-17)
+
+
+### Features
+
+* add sass-loader options ([76fbfba](https://github.com/ykfe/ssr/commit/76fbfbaada67fda444954686ce10f416e46766db))
+
+
+
+
+
+## [5.5.49](https://github.com/ykfe/ssr/compare/v5.5.48...v5.5.49) (2021-06-10)
+
+**Note:** Version bump only for package ssr-types
+
+
+
+
+
+## [5.5.46](https://github.com/ykfe/ssr/compare/v5.5.45...v5.5.46) (2021-06-02)
+
+**Note:** Version bump only for package ssr-types
+
+
+
+
+
+## [5.5.45](https://github.com/ykfe/ssr/compare/v5.5.44...v5.5.45) (2021-06-02)
+
+
+### Features
+
+* add babelExtraModule ([ff47710](https://github.com/ykfe/ssr/commit/ff47710c5cee4fff14487394f052295b8f564459))
+
+
+
+
+
+## [5.5.44](https://github.com/ykfe/ssr/compare/v5.5.43...v5.5.44) (2021-06-02)
+
+**Note:** Version bump only for package ssr-types
+
+
+
+
+
+## [5.5.32](https://github.com/ykfe/ssr/compare/v5.5.31...v5.5.32) (2021-05-20)
+
+**Note:** Version bump only for package ssr-types
+
+
+
+
+
+## [5.5.30](https://github.com/ykfe/ssr/compare/v5.5.29...v5.5.30) (2021-05-20)
+
+
+### Features
+
+* 支持 https ([e91808c](https://github.com/ykfe/ssr/commit/e91808c91cb2b8860f0771518b48de4dd592fcda))
+
+
+
+
+
+## [5.5.28](https://github.com/ykfe/ssr/compare/v5.5.27...v5.5.28) (2021-05-13)
+
+**Note:** Version bump only for package ssr-types
+
+
+
+
+
+## [5.5.26](https://github.com/ykfe/ssr/compare/v5.5.25...v5.5.26) (2021-05-12)
+
+
+### Features
+
+* react 场景支持 basename ([1cd6e94](https://github.com/ykfe/ssr/commit/1cd6e94873c4ed75c7f06868345afe93004546b0))
+
+
+
+
+
+## [5.5.23](https://github.com/ykfe/ssr/compare/v5.5.22...v5.5.23) (2021-05-10)
+
+
+### Features
+
+* add ssrVueLoaderOptions csrVueloaderOptions ssr-hoc-vue3 ([09c4dc5](https://github.com/ykfe/ssr/commit/09c4dc5a820e4c57ece73427a78968b75abea2f3))
+
+
+
+
+
+## [5.5.23](https://github.com/ykfe/ssr/compare/v5.5.22...v5.5.23) (2021-05-10)
+
+
+### Features
+
+* add ssrVueLoaderOptions csrVueloaderOptions ssr-hoc-vue3 ([9ad8187](https://github.com/ykfe/ssr/commit/9ad81877b8cc35b8d98f6c8b3cc62e8071ed2dcf))
