+++ packages/types-react/package.json
@@ -1,10 +1,9 @@
 {
-  "name": "ssr-hoc-react",
+  "name": "ssr-types-react",
   "version": "5.5.57",
-  "description": "hoc for react ssr",
+  "description": "types for react ssr",
   "main": "./cjs/index.js",
   "module": "./esm/index.js",
-  "sideEffects": false,
   "scripts": {
     "build": "concurrently \"tsc -p ./tsconfig.cjs.json --inlineSourceMap false\" \" tsc -p ./tsconfig.esm.json --inlineSourceMap false\"",
     "build:only": "concurrently \"tsc-transpile-only --skipLibCheck -p ./tsconfig.cjs.json --inlineSourceMap false\" \" tsc-transpile-only --skipLibCheck -p ./tsconfig.esm.json --inlineSourceMap false\"",
@@ -29,14 +28,12 @@
   },
   "homepage": "https://github.com/ykfe/ssr#readme",
   "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.0",
-    "react-router-dom": "^5.1.2"
-  },
-  "devDependencies": {
+    "ssr-types": "^5.5.57",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.3",
-    "ssr-types-react": "^5.5.57"
+    "@types/react-router-dom": "^5.1.6"
+  },
+  "devDependencies": {
+    "concurrently": "^5.1.0"
   }
 }
