+++ packages/types-react/src/component.ts
