+++ packages/types/package.json
@@ -1,7 +1,7 @@
 {
   "name": "ssr-types",
   "version": "5.5.57",
-  "description": "cli for serverless ssr",
+  "description": "types for serverless ssr",
   "main": "./cjs/index.js",
   "module": "./esm/index.js",
   "scripts": {
@@ -31,9 +31,6 @@
     "@midwayjs/faas-typings": "^2.7.0",
     "@types/express": "^4.17.11",
     "@types/koa": "^2.11.3",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.6",
     "@types/tapable": "^1.0.6",
     "@types/webpack": "^4.41.10",
     "@types/yargs": "^13.0.4",
