+++ packages/types/src/index.ts
@@ -1,7 +1,10 @@
 export * from './ctx'
-export * from './route'
 export * from './config'
-export * from './component'
 export * from './yargs'
 
 export type Mode = 'development' |'production'
+
+export type ESMFeRouteItem<T={}> = {
+  path: string
+  webpackChunkName: string
+} & T
